Right Click/requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
opencv-python
torch